File sizes stored in the hash table can be larger than integers.

+++ mirrordoctor/mb/conn.py
@@ -181,7 +181,7 @@ class Conn:
             CREATE TABLE "hash" (
                     "file_id" INTEGER REFERENCES filearr PRIMARY KEY,
                     "mtime" INTEGER NOT NULL,
-                    "size"  INTEGER NOT NULL,
+                    "size" BIGINT NOT NULL,
                     "md5"    BYTEA NOT NULL,
                     "sha1"   BYTEA NOT NULL,
                     "sha256" BYTEA NOT NULL,
